feat: API 함수 타입스크립트 변환
